Fix some more db errors with cards
