+++ routes/account.js
@@ -11,4 +11,22 @@ router.get("/", async(req, res) => {
     }
 });
 
+router.get("/decks", async(req, res) => {
+    try {
+        let results = {};
+        let deckList = await decks.getDecksByOwner(req.params.userId);
+
+        results["decks"] = deckList;
+
+        if (req.user) {
+            let user = await users.getUser(req.user._id);
+            results["username"] = user.username;
+        }
+
+        res.render("decks/index", results);
+    } catch (err) {
+        res.status(404).json({message: "Something went wrong"});
+    }
+});
+
 module.exports = router;
