+++ routes/decks.js
@@ -23,27 +23,8 @@ router.get("/", async(req, res) => {
     }
 });
 
-router.get("/:userId", async(req, res) => {
-    try {
-        let results = {};
-        let deckList = await decks.getDecksByOwner(req.params.userId);
-
-        results["decks"] = deckList;
-
-        if (req.user) {
-            let user = await users.getUser(req.user._id);
-            results["username"] = user.username;
-        }
-
-        res.render("decks/index", results);
-    } catch (err) {
-        res.status(404).json({message: "Something went wrong"});
-    }
-});
-
 
 router.get("/:deckId", async(req, res) => {
-
     try {
         let results = {};
         let deck = await decks.getDeck(req.params.deckId);
@@ -52,8 +33,10 @@ router.get("/:deckId", async(req, res) => {
 
         if (req.user) {
             let user = await users.getUser(req.user._id);
-            results["username"] = user.username;
+            results.username = user.username;
         }
+
+        console.log(results);
 
         res.render("decks/instance", results);
     } catch (err) {
